Remove negativeCounterAllowed
Remove all config and data attributes for controlling negative counts.  This was removed because the library can't guarantee a non-negative count 100% of the time, and it's dubious that this is requirement of a sharded-counter.  If non-negative counters are desired, the clients can implement this guard in their code, or rely upon a different type of counter in this library (yet to come)